Create packages config and models
